administator

+++ WebApplication2/Controllers/AdminController.cs
@@ -4,23 +4,31 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using System.Data.Entity;
+
 using WebApplication2.Models;
+using Microsoft.AspNet.Identity.Owin;
 
 namespace WebApplication2.Controllers
 {
     public class AdminController : Controller
     {
+        UserContext db = new UserContext();
+
         public ActionResult Admin()
+        {
+            var UserLogin = User.Identity.Name;
+            ViewBag.UserLogin = UserLogin;
+            return View();
+        }
+
+        [Authorize(Roles = "admin")]
+        public ActionResult ViewAdmin()
         {
             try
             {
-                using (UserContext db = new UserContext())
-                {
-                    IEnumerable<User> users = db.Users.ToList();
-                    ViewBag.UserInfo = users;
-                    return View();
-                }
-                 
+                IEnumerable<User> users = db.Users.ToList();
+                ViewBag.UserInfo = users;
+                return View();
             }
             catch (Exception ex)
             {
@@ -29,23 +37,49 @@ namespace WebApplication2.Controllers
         }
 
         [HttpPost]
-        public ActionResult Edit_User(string action,int Id)
+        [Authorize(Roles = "admin")]
+        public ActionResult EditUser(string product, string action, int Id)
         {
+            if (action == "Edit")
+            {
+                return RedirectToAction("UpdateUser", "Admin",new { @Id=Id});
+            }
             if (action == "Delete")
             {
-                using(UserContext db=new UserContext())
+                User u = db.Users.Find(Id);
+                if (u != null)
                 {
-                    User u = db.Users.Find(Id);
-                    if (u != null)
-                    {
-                        db.Users.Remove(u);
-                        db.SaveChanges();
-                    }
-                    return RedirectToAction("Admin", "Admin");
+                    db.Users.Remove(u);
+                    db.SaveChanges();
                 }
-               
+                return RedirectToAction("ViewAdmin", "Admin");
             }
-            return View();
+
+            return HttpNotFound();
+        }
+
+        [HttpGet]
+        public ActionResult UpdateUser(int? Id)
+        {
+            if (Id == null)
+            {
+                return HttpNotFound();
+            }
+            User user = db.Users.Find(Id);
+
+            if (user == null)
+            {
+                return HttpNotFound();
+            }
+            return View(user);
+        }
+
+        [HttpPost]
+        public ActionResult UpdateUser(User user)
+        {
+            db.Entry(user).State = EntityState.Modified;
+            db.SaveChanges();
+            return RedirectToAction("ViewAdmin", "Admin");
         }
     }
 }
